api-gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
description = ""
authors = [
    {name = "Matias Vagliviello",email = "vaglimatias@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.120.2,<0.121.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "PyJWT (>=2.7.0,<3.0.0)"
]


[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=2.0.0,<3.0.0"]
package-mode = false

[tool.poetry]
name = "api-gateway"
version = "0.1.0"
description = ""
authors = ["Matias"]
package-mode = false
